+++ python/lsst/ts/fbs/utils/auxtel/surveys.py
@@ -20,19 +20,44 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 import typing
-
+from ..utils import get_data_dir
 import astropy.units as u
 import numpy as np
-from rubin_scheduler.scheduler.detailers import (BaseDetailer,
-                                                 TrackingInfoDetailer)
-from rubin_scheduler.scheduler.surveys import (BaseSurvey, FieldSurvey,
-                                               GreedySurvey)
+import yaml
+from rubin_scheduler.scheduler.detailers import BaseDetailer, TrackingInfoDetailer
+from rubin_scheduler.scheduler.surveys import BaseSurvey, FieldSurvey, GreedySurvey
 from rubin_scheduler.scheduler.utils import ObservationArray
 
 from ..target import Target
-from .basis_functions import (get_basis_functions_cwfs_survey,
-                              get_basis_functions_image_survey,
-                              get_basis_functions_spectroscopic_survey)
+from .basis_functions import (
+    get_basis_functions_cwfs_survey,
+    get_basis_functions_image_survey,
+    get_basis_functions_spectroscopic_survey,
+)
+
+
+def get_auxtel_targets(infile: str | None = None) -> dict:
+    """Load potential targets for auxtel observations.
+
+    Targets can be split into different categories for convenience.
+    The targets in a particular category should be destined for the same
+    'kind' of survey (imaging survey or spectroscopy survey), so that they
+    can be configured similarly.
+
+    After target_dict is returned, remove or include desired targets.
+
+    Returns
+    -------
+    target_pointings : `dict`
+        Dictionary of candidate target information.
+        The minimum necessary is ra, dec, and block.
+        Additional options include exptime, nexp, visit_gap, and priority.
+    """
+    if infile is None:
+        infile = get_data_dir() / "auxtel_targets.yaml"
+    with open(infile) as stream:
+        target_pointings = yaml.safe_load(stream)
+    return target_pointings
 
 
 def generate_image_survey_from_tiles(
@@ -67,8 +92,8 @@ def generate_image_survey_from_tiles(
     basis_functions = get_basis_functions_image_survey(
         ra=target.ra.to(unit=u.deg).value,
         nside=nside,
-        note=target.target_name,
-        note_interest=target.survey_name,
+        note=target.survey_name,
+        note_interest=target.science_program,
         ha_limits=target.hour_angle_limit,
         wind_speed_maximum=wind_speed_maximum,
         nobs_reference=nfields,
@@ -79,19 +104,14 @@ def generate_image_survey_from_tiles(
     )
 
     # Set up a sequence of nexp exposures per filter
-    sequence_len = target.nexp * len(target.filters)
-    sequence = ObservationArray(n=sequence_len)
-    sequence["RA"] = target.ra.to(u.rad).value
-    sequence["dec"] = target.dec.to(u.rad).value
-    sequence["exptime"] = target.exptime
-    sequence["nexp"] = 1
-    sequence["scheduler_note"] = f"{target.survey_name}:{target.target_name}"
-    sequence["target_name"] = f"{target.target_name}"
-    sequence["science_program"] = f"{target.survey_name}"
-    i = 0
-    for filter_obs in target.filters:
-        sequence["filter"][i : i + target.nexp] = filter_obs
-        i += target.nexp
+    sequence = [ObservationArray(n=1) for i in range(len(target.filters))]
+    for filter_obs, observation in zip(target.filters, sequence):
+        observation["RA"] = target.ra.to(u.rad).value
+        observation["dec"] = target.dec.to(u.rad).value
+        observation["filter"] = filter_obs
+        observation["exptime"] = target.exptime
+        observation["nexp"] = target.nexp
+        observation["scheduler_note"] = target.survey_name
 
     image_survey = FieldSurvey(
         basis_functions,
@@ -108,16 +128,10 @@ def generate_image_survey_from_tiles(
         sequence=sequence,
         survey_name=f"{target.survey_name}",
         target_name=f"{target.target_name}",
-        science_program=f"{target.survey_name}",
-        scheduler_note=f"{target.survey_name}:{target.target_name}",
+        science_program=f"{target.science_program}",
+        scheduler_note=f"{target.survey_name}",
         nside=nside,
-        detailers=[
-            TrackingInfoDetailer(
-                target_name=f"{target.target_name}",
-                science_program=f"{target.survey_name}",
-            )
-        ]
-        + survey_detailers,
+        detailers=survey_detailers,
     )
     # Weight all of the basis functions up or down to weight overall survey
     image_survey.basis_weights *= target.reward_value
@@ -130,6 +144,8 @@ def generate_image_survey_from_target(
     target: Target,
     wind_speed_maximum: float,
     survey_detailers: typing.List[BaseDetailer],
+    avoid_wind: bool = True,
+    include_slew: bool = True,
 ) -> BaseSurvey:
     """Generate image survey, for single Target with dithers
 
@@ -143,6 +159,12 @@ def generate_image_survey_from_target(
         Wind speed limit, in m/s.
     survey_detailers : `list` of `detailers.BaseDetailer`
         List of survey detailers.
+    avoid_wind : `bool`, optional
+        If True, add the wind avoidance basis function.
+        If False, drop basis function entirely.
+        Makes use align with the spectroscopic survey.
+    include_slew : `bool`, optional
+        If True, include slewtime basis functions.
 
     Returns
     -------
@@ -162,31 +184,27 @@ def generate_image_survey_from_target(
         filter_names=target.filters,
         gap_min=target.visit_gap,
         additional_notes=None,
+        include_slew=include_slew,
     )
 
     # Set up a sequence of nexp exposures per filter
-    sequence_len = target.nexp * len(target.filters)
-    sequence = ObservationArray(n=sequence_len)
-    sequence["RA"] = target.ra.to(u.rad).value
-    sequence["dec"] = target.dec.to(u.rad).value
-    sequence["exptime"] = target.exptime
-    sequence["nexp"] = 1
-    sequence["scheduler_note"] = f"{target.survey_name}:{target.target_name}"
-    sequence["target_name"] = f"{target.target_name}"
-    sequence["science_program"] = f"{target.survey_name}"
-    i = 0
-    for filter_obs in target.filters:
-        sequence["filter"][i : i + target.nexp] = filter_obs
-        i += target.nexp
+    sequence = [ObservationArray(n=target.nexp) for i in range(len(target.filters))]
+    for filter_obs, observation in zip(target.filters, sequence):
+        observation["RA"] = target.ra.to(u.rad).value
+        observation["dec"] = target.dec.to(u.rad).value
+        observation["filter"] = filter_obs
+        observation["exptime"] = target.exptime
+        observation["nexp"] = 1
+        observation["scheduler_note"] = target.survey_name
 
     image_survey = FieldSurvey(
-        basis_functions,
-        np.array(
+        basis_functions=basis_functions,
+        RA=np.array(
             [
                 target.ra.to(u.degree).value,
             ]
         ),
-        np.array(
+        dec=np.array(
             [
                 target.dec.to(u.degree).value,
             ]
@@ -194,16 +212,10 @@ def generate_image_survey_from_target(
         sequence=sequence,
         survey_name=f"{target.survey_name}",
         target_name=f"{target.target_name}",
-        science_program=f"{target.survey_name}",
-        scheduler_note=f"{target.survey_name}:{target.target_name}",
+        science_program=f"{target.science_program}",
+        scheduler_note=f"{target.survey_name}",
         nside=nside,
-        detailers=[
-            TrackingInfoDetailer(
-                target_name=f"{target.target_name}",
-                science_program=f"{target.survey_name}",
-            )
-        ]
-        + survey_detailers,
+        detailers=survey_detailers,
     )
     # Weight all of the basis functions up or down to weight overall survey
     image_survey.basis_weights *= target.reward_value
@@ -244,9 +256,6 @@ def generate_cwfs_survey(
 
     return GreedySurvey(
         basis_functions,
-        # Boosting the reward very high is good if this is competing
-        # with other surveys in a tier, but otherwise it just blows up
-        # plots of reward value over time
         np.ones_like(basis_functions) * 1.0,
         nside=nside,
         survey_name="cwfs",
@@ -262,6 +271,7 @@ def generate_spectroscopic_survey(
     wind_speed_maximum: float,
     nfields: int,
     survey_detailers: typing.List[BaseDetailer],
+    include_slew: bool = True,
 ) -> BaseSurvey:
     """Generate Spectroscopic Survey.
 
@@ -277,6 +287,8 @@ def generate_spectroscopic_survey(
         Maximum wind speed (in m/s).
     survey_detailers : `list` of `detailers.BaseDetailer`
         List of survey detailers.
+    include_slew : `bool`, optional
+        If True, include slewtime basis functions.
 
     Returns
     -------
@@ -294,18 +306,17 @@ def generate_spectroscopic_survey(
         gap_min=target.visit_gap,
         moon_distance=target.moon_distance,
         nobs_reference=nfields,
+        include_slew=include_slew,
     )
 
-    observation = ObservationArray(n=1)
+    observation = ObservationArray(n=target.nexp)
     observation["RA"] = target.ra.to(u.rad).value
     observation["dec"] = target.dec.to(u.rad).value
-    # This can change to band once FBS updated to 3.5+
     observation["filter"] = "r"
     observation["exptime"] = target.exptime
-    observation["nexp"] = target.nexp
-    observation["scheduler_note"] = f"{target.survey_name}:{target.target_name}"
-    observation["target_name"] = f"{target.target_name}"
-    observation["science_program"] = f"{target.survey_name}"
+    observation["nexp"] = 1
+    observation["scheduler_note"] = target.survey_name
+    observation = [observation]
 
     spectroscopic_survey = FieldSurvey(
         basis_functions,
@@ -320,18 +331,12 @@ def generate_spectroscopic_survey(
             ]
         ),
         sequence=observation,
-        survey_name=f"{target.survey_name}",  # This is target name on purpose.
+        survey_name=f"{target.survey_name}",
         target_name=f"{target.target_name}",
-        scheduler_note=f"{target.target_name}",
-        science_program=f"{target.survey_name}",
+        scheduler_note=f"{target.survey_name}",
+        science_program=f"{target.science_program}",
         nside=nside,
-        detailers=[
-            TrackingInfoDetailer(
-                target_name=f"{target.target_name}",
-                science_program=f"{target.survey_name}",
-            )
-        ]
-        + survey_detailers,
+        detailers=survey_detailers,
     )
     # Weight all of the basis functions up or down to weight overall survey
     spectroscopic_survey.basis_weights *= target.reward_value