+++ tests/auxtel/test_basis_functions.py
@@ -39,9 +39,10 @@ def test_get_basis_functions_image_survey() -> None:
         filter_names=["g", "r", "i"],
         gap_min=144.0,
         additional_notes=[["unit_test", 32]],
+        include_slew=True,
     )
 
-    assert len(basis_functions) == 12
+    assert len(basis_functions) == 11
 
 
 def test_get_basis_functions_cwfs_survey() -> None:
@@ -52,7 +53,7 @@ def test_get_basis_functions_cwfs_survey() -> None:
         wind_speed_maximum=7.0,
     )
 
-    assert len(basis_functions) == 9
+    assert len(basis_functions) == 8
 
 
 def test_get_basis_functions_spectroscopic_survey() -> None:
@@ -69,7 +70,7 @@ def test_get_basis_functions_spectroscopic_survey() -> None:
         nobs_reference=3,
     )
 
-    assert len(basis_functions) == 10
+    assert len(basis_functions) == 8
 
 
 def test_get_basis_functions_spectroscopic_survey_no_wind() -> None:
@@ -86,4 +87,4 @@ def test_get_basis_functions_spectroscopic_survey_no_wind() -> None:
         nobs_reference=3,
     )
 
-    assert len(basis_functions) == 9
+    assert len(basis_functions) == 7